+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `reproject_shape` without a precision ([#454](https://github.com/stac-utils/stactools/pull/454))
+
 ## [0.5.0] - 2023-08-04
 
 ### Added